+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
+## [3.7.2](https://github.com/stellar/stellar-disbursement-platform-backend/releases/tag/3.7.2) ([diff](https://github.com/stellar/stellar-disbursement-platform-backend/compare/3.7.1...3.7.2))
+
+### Added
+- Added Receiver's External ID to the `GET /exports/payments` endpoint. [#748](https://github.com/stellar/stellar-disbursement-platform-backend/pull/748)
+
 ## [3.7.1](https://github.com/stellar/stellar-disbursement-platform-backend/releases/tag/3.7.1) ([diff](https://github.com/stellar/stellar-disbursement-platform-backend/compare/3.7.0...3.7.1))
 
 ### Fixed